Refactors ListParticipantsAdmin
